updated tests/test_unix.sh

+++ tests/test_unix.sh
@@ -38,6 +38,11 @@ $perl -T $I_lib ./check_file_md5.pl      -f test.txt -v -c 'd8e8fca2dc0f896fd7cb
 $perl -T $I_lib ./check_file_sha1.pl     -f test.txt -v --checksum '4e1243bd22c66e76c2ba9eddc1f91394e57f9f83'
 $perl -T $I_lib ./check_file_sha256.pl   -f test.txt -v -c 'f2ca1bb6c7e907d06dafe4687e579fce76b37e4e93b7605022da52e6ccc26fd2'
 $perl -T $I_lib ./check_file_sha512.pl   -f test.txt -v -c '0e3e75234abc68f4378a86b3f4b32a198ba301845b0cd6e50106e874345700cc6663a86c1ea125dc5e92be17c98f9a0f85ca9d5f595db2012f7cc3571945c123'
+echo "Testing check_ssh_login.pl correctly fails on non-existent user against localhost"
+set +e
+$perl -T $I_lib ./check_ssh_login.pl -H localhost -u check_ssh_login_nagios_plugin_test -p test
+[ $? -eq 2 ] || exit 1
+set -e
 rm -f test.txt
 hr
 echo; echo
